US24682 Add check for registration before upload
Former-commit-id: 48acf652c41a9094f7b3a2d0c5ca7260439ba09b

+++ hpz/hpz/frs/upload_service.py
@@ -1,6 +1,7 @@
 """
 This module describes the file upload endpoint for HPZ.
 """
+import logging
 
 __author__ = 'ablum,'
 __author__ = 'tshewchuk'
@@ -10,6 +11,8 @@ from pyramid.view import view_config
 import os
 from hpz.database.file_registry import FileRegistry
 
+logger = logging.getLogger(__name__)
+
 
 @view_config(route_name='files', renderer='json', request_method='POST')
 def file_upload_service(context, request):
@@ -17,19 +20,24 @@ def file_upload_service(context, request):
     registration_id = request.matchdict['registration_id']
     file_name = request.headers['Filename']
     base_upload_path = request.registry.settings['hpz.frs.upload_base_path']
-
     file_pathname = os.path.join(base_upload_path, registration_id + file_name)
 
-    for item, f in request.POST.items():
-        with open(file_pathname, mode='wb',) as new_file:
-            input_file = f.file
-            input_file.seek(0)
-            while True:
-                data = input_file.read(2 << 16)
-                if not data:
-                    break
-                new_file.write(data)
+    if FileRegistry.is_file_registered(registration_id):
 
-    FileRegistry.file_upload_request(registration_id, file_pathname)
+        for item, f in request.POST.items():
+            with open(file_pathname, mode='wb',) as new_file:
+                input_file = f.file
+                input_file.seek(0)
+                while True:
+                    data = input_file.read(2 << 16)
+                    if not data:
+                        break
+                    new_file.write(data)
+
+        FileRegistry.file_upload_request(registration_id, file_pathname)
+        logger.error('This file was successfully uploaded')
+
+    else:
+        logger.error('The file attempting to be upload is not registered')
 
     return Response()
